+++ modules/sde-submodules/assembly-part-relationship/src/main/java/org/eclipse/tractusx/sde/submodels/apr/model/AspectRelationship.java
@@ -22,9 +22,7 @@
 
 package org.eclipse.tractusx.sde.submodels.apr.model;
 
-import java.util.List;
-
-import org.eclipse.tractusx.sde.common.entities.Policies;
+import org.eclipse.tractusx.sde.common.entities.CommonPropEntity;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -32,39 +30,18 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 @Builder
 @Data
+@EqualsAndHashCode(callSuper=false)
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class AspectRelationship {
+public class AspectRelationship extends CommonPropEntity {
 
-	private String shellId;
-	private String subModelId;
-	private String usagePolicyId;
-	private String assetId;
-	private String accessPolicyId;
-	private String contractDefinationId;
-	private String deleted;
-	private String updated;
 	private String oldSubmodelIdforUpdateCase;
-
-	@JsonProperty(value = "row_number")
-	private Integer rowNumber;
-
-	@JsonProperty(value = "process_id")
-	private String processId;
-
-	@JsonProperty(value = "bpn_numbers")
-	private List<String> bpnNumbers;
-
-	@JsonProperty(value = "type_of_access")
-	private String typeOfAccess;
-
-	@JsonProperty(value = "usage_policies")
-	private List<Policies> usagePolicies;
 
 	@JsonProperty(value = "uuid")
 	private String childUuid;