+++ modules/sde-submodules/part-as-planned/src/main/java/org/eclipse/tractusx/sde/submodels/pap/steps/EDCPartAsPlannedHandlerStep.java
@@ -19,10 +19,10 @@
  ********************************************************************************/
 package org.eclipse.tractusx.sde.submodels.pap.steps;
 
-import java.util.List;
 import java.util.Map;
 
 import org.eclipse.tractusx.sde.common.constants.CommonConstants;
+import org.eclipse.tractusx.sde.common.entities.PolicyModel;
 import org.eclipse.tractusx.sde.common.exception.CsvHandlerUseCaseException;
 import org.eclipse.tractusx.sde.common.exception.ServiceException;
 import org.eclipse.tractusx.sde.common.submodel.executor.Step;
@@ -48,7 +48,7 @@ public class EDCPartAsPlannedHandlerStep extends Step {
 	private final PartAsPlannedService partAsPlannedService;
 
 	@SneakyThrows
-	public PartAsPlanned run(String submodel, PartAsPlanned input, String processId) {
+	public PartAsPlanned run(String submodel, PartAsPlanned input, String processId, PolicyModel policy) {
 		String shellId = input.getShellId();
 		String subModelId = input.getSubModelId();
 
@@ -56,13 +56,12 @@ public class EDCPartAsPlannedHandlerStep extends Step {
 
 			AssetEntryRequest assetEntryRequest = assetFactory.getAssetRequest(submodel,
 					getSubmodelShortDescriptionOfModel(), shellId, subModelId, input.getUuid());
-			if (!edcGateway.assetExistsLookup(
-					assetEntryRequest.getId())) {
-				edcProcessingforPartAsPlanned(assetEntryRequest, input);
+			if (!edcGateway.assetExistsLookup(assetEntryRequest.getId())) {
+				edcProcessingforPartAsPlanned(assetEntryRequest, input, policy);
 			} else {
 
 				deleteEDCFirstForUpdate(submodel, input, processId);
-				edcProcessingforPartAsPlanned(assetEntryRequest, input);
+				edcProcessingforPartAsPlanned(assetEntryRequest, input, policy);
 				input.setUpdated(CommonConstants.UPDATED_Y);
 			}
 
@@ -86,10 +85,10 @@ public class EDCPartAsPlannedHandlerStep extends Step {
 	}
 
 	@SneakyThrows
-	private void edcProcessingforPartAsPlanned(AssetEntryRequest assetEntryRequest, PartAsPlanned input) {
+	private void edcProcessingforPartAsPlanned(AssetEntryRequest assetEntryRequest, PartAsPlanned input,
+			PolicyModel policy) {
 
-		Map<String, String> createEDCAsset = createEDCAssetFacilator.createEDCAsset(assetEntryRequest,
-				List.of(), List.of());
+		Map<String, String> createEDCAsset = createEDCAssetFacilator.createEDCAsset(assetEntryRequest, policy);
 
 		// EDC transaction information for DB
 		input.setAssetId(assetEntryRequest.getId());