+++ modules/sde-submodules/pcf/src/main/java/org/eclipse/tractusx/sde/submodels/pcf/repository/PcfRepository.java
@@ -1,6 +1,6 @@
 /********************************************************************************
- * Copyright (c) 2023 T-Systems International GmbH
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023, 2024 T-Systems International GmbH
+ * Copyright (c) 2023, 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -21,6 +21,7 @@
 package org.eclipse.tractusx.sde.submodels.pcf.repository;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.eclipse.tractusx.sde.submodels.pcf.entity.PcfEntity;
 import org.springframework.data.jpa.repository.Query;
@@ -28,11 +29,11 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface PcfRepository extends CrudRepository<PcfEntity, String> {
 
-	List<PcfEntity> findByProcessIdforPcf(String processIdforPcf);
+	List<PcfEntity> findByProcessId(String processId);
 
-	@Query("select count(ae) from PcfEntity ae where ae.updatedforPcf = ?1 and ae.processIdforPcf = ?2")
+	@Query("select count(ae) from PcfEntity ae where ae.updated = ?1 and ae.processId = ?2")
 	long countByUpdatedAndProcessId(String updated, String processId);
-	
-	
+
+	Optional<PcfEntity> findByProductId(String productId);
 
 }